+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - add field `fields_map` in style `showUserInput`; Loads the defined fields from the mapper in the table visualization.
  - remove fields `submit_and_send_label`, `email_body`, `submit_and_send_email`, `email_address` and `email_subject` from form styles. This functionality can be achieved with actions
  - remove style `emailForm`
+ - add new style `entryRecordDelete`
 
 ### Bugfix
  - set values for `JSON` content fields properly in CMS